[refactoring] 1. Car 생성 로직 RacingCars로 이동 2. 구현체 생성 조금 더 고민 ,,
